requirements.txt
torch>=2.0.1
torchaudio>=2.0.2
whisperx>=3.1.0
pyannote.audio>=3.0.0
numpy>=1.21.0
tqdm
python-dotenv